Move tcp_create_socket impl
