+++ crates/wasi/src/preview2/host/mod.rs
@@ -7,6 +7,5 @@ mod io;
 pub(crate) mod network;
 mod random;
 mod tcp;
-mod tcp_create_socket;
 mod udp;
 mod udp_create_socket;
